edge_device/requirements.txt
psutil>=7.0.0
PyYAML>=6.0.2
requests>=2.31.3
pytest>=8.4.0